% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsar-package.R
\docType{package}
\name{atsar-package}
\alias{atsar-package}
\alias{atsar}
\title{The 'atsar' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
