% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dfa.R
\name{fit_dfa}
\alias{fit_dfa}
\title{fit_dfa is the primary function for fitting dynamic factor analysis to time series data.}
\usage{
fit_dfa(y = y, covar = NULL, covar_index = NULL, num_trends = 2,
  varIndx = NULL, zscore = TRUE, iter = 4000, chains = 1,
  control = list(adapt_delta = 0.99))
}
\arguments{
\item{y}{The response variable (numeric matrix)}

\item{covar}{The optional matrix of covariates (time on columns)}

\item{covar_index}{The matrix of index values, indicating which covariate effects on time series are shared.}

\item{num_trends}{The number of DFA trends (random walks) to estimate. Process variance fixed at 1.}

\item{varIndx}{The vector indicating which time series share variances. By default, (1,2,3,...,n)}

\item{zscore}{Whether or not to z-score data. Defaults to true.}

\item{iter}{Number of MCMC iterations, defaults to 4000.}

\item{chains}{Number of MCMC chains, defaults to 1 (for DFA this is the best choice)}

\item{control}{Default control list for stanfit objects.}
}
\value{
an object of class 'rstan'
}
\description{
fit_dfa is the primary function for fitting dynamic factor analysis to time series data.
}
