% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_trends.R
\name{rotate_trends}
\alias{rotate_trends}
\title{rotate_trends is the primary function which calls pre-written stan scripts for time series data.}
\usage{
rotate_trends(fitted_model)
}
\arguments{
\item{fitted_model}{A fitted stanfit object}
}
\value{
a list object, with the rotated trends for each MCMC chain (trends), rotated Z for every MCMC chain (Z_rot), mean Z (Z_rot_mean), mean trends (trends_mean), lower 2.5% interval on trends (trends_lower), upper 97.5% interval for trends (trends_upper)
}
\description{
rotate_trends is the primary function which calls pre-written stan scripts for time series data.
}
